CHROMIUM: usb: gadget: Fix handling of unrecognized OS descriptor requests
If composite_setup() receives a control request for an OS Feature
Descriptor that is not Extended Compatibility ID or Extended Properties,
it falls through to enqueueing a response of length 'value' bytes which,
because the request is not recognized, is -EOPNOTSUPP.  Fix it so that
we instead bail out of composite_setup() without enqueueing a response.
The UDC driver is then responsible for halting EP0.

BUG=chrome-os-partner:47809
TEST=Build and boot on Smaug.

Change-Id: I363c53a35a45af5cb571370ae18d098c2f169de5
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313440