CHROMIUM: drm/tegra: Always boot nvdec with lower clk frequency
nvdec does not boot reliably when pll_c3 is set to max 716.8MHz. This
change is to lower default rate, the clock takes when it is first
enabled. Usually when a clock is first enabled, it takes the last
clock frequency that it was set to. With change "CHROMIUM: host1x:
Implement per-user clock constraints", nvdec.cbus and inturn pll_c3
clk rate is set to max once disabled, so when enabled, clock rate
is set to 716.8MHz and this is causing intermittent boot failures
for nvdec falcon.

BUG=chrome-os-partner:47453
TEST=run youtube, put device to LP0, resume from LP0 and repeat the
     cycle, should not observe nvdec boot failure or nvdec syncpt
	 timeout.

Change-Id: Ia11cbab28ee6f415787549c4f1a9f4287ef63663
Signed-off-by: Maneet Singh <mmaneetsingh@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/313501
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>