CHROMIUM: usb: gadget: tegra_xudc: Properly set PORTPM L1S
PortPM L1S bit is used to accept or reject L1 entry request.  When LPM
is disabled, PortPM L1S should be set to NYET, and when LPM is enabled,
we should also set the L1S to ACCEPT.

This is optional for now since lpm_enable value cannot be changed at
run time, but should be done if lpm_enable is configurable at run time
in the future.

BUG=none
TEST=Build and boot on Smaug

Change-Id: If86727ea2395669ef322f577afb3eb34c463b75b
Signed-off-by: George Zhou <gzhou@nvidia.com>
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/313856
Commit-Ready: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>