Revert "CHROMIUM: atmel_mxt_ts: sleep after issuing reset command"
This reverts commit 137b5e89471d ("CHROMIUM: atmel_mxt_ts: sleep after
issuing reset command") because it does not allow updating firmware on
Celes: we always time out when entering bootloader mode because we miss
CHG transition.

I think we need to come up with better strategy of validating entering
bootloader mode that also works on Samus, probably by trying to read
bootloader status in interrupt before signalling bl_completion, but
that's for a followup change.

BUG=chrome-os-partner:44795
TEST=run touchpad firmware update on celes and verify that it completes
     without errors:
     chromeos-touch-firmware-update.sh -d ATML0000:00 -n maxtouch-tp.fw -f

Change-Id: I36d93a838ffecc8459ca8ffa0316156b1f067738
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/296914
Reviewed-by: Benson Leung <bleung@chromium.org>