CHROMIUM: drm/nouveau: Prevent page size promotion on remap
When remapping a buffer into an address space allocation, if the buffer
is using large page size but the address space allocation allows things
that are mapped into it to use small page size, honor this by demoting the
buffer to using small page size (which was being done properly).  However,
the reverse is not valid...if the buffer was originally using small page
size (for presumably valid reasons), and the address space allocation
allows things that are mapped into it to use large page size, we cannot
promote that buffer to be mapped using large page size, so prevent that
from happening.

BUG=chrome-os-partner:42636
TEST=Built and boot kernel (actually tested on branch
stabilize-smaug-7566.B-chromeos-3.18).  Ran internal compute test that ran
into this case (causing a fixed address remap failure) and make sure it
passes.

Change-Id: Icd2e357eaefffbe5e1e70abcdea33d15191e0440
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/314835
Reviewed-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>