CHROMIUM: ASoC: rt5677: Keep a wake lock for hotword stream
Right after a hotword irq finishes resuming the system, sometimes
a suspend is followed immediately, freezing the delayed work item
which is reading the DSP mic buffer. The system sometimes also
suspends, half way during a DSP fw load. From the user's point of
view, the device didn't responds to the hotword in both cases.

This patch adds a wake_lock to keep the system awake while
loading the DSP fw and while streaming audio from the DSP mic buffer.
The lock is taken at pcm open and also in the hotword irq handler.
The lock is released when the DSP is armed or if userspace closes
the hotword pcm.

BUG=chrome-os-partner:48039
TEST=Ryu always responds to hotword during suspend

Change-Id: I1c395f1d50244b5abd0a463a1ee9f914a0cc3291
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/316924
Reviewed-by: Dylan Reid <dgreid@chromium.org>