Revert "FROMLIST: PM / Domains: add setter for dev.pm_domain"
This reverts commit 8a679b90b2130dd3970ed41f3a18df6030b71b1c.

This patch conflicts with many upstream patches that we would
like to land in chromeos-3.18.
Furthermore, this is a FROMLIST patch that is still under review
on the list for a feature that is not yet enabled.
So, let's just revert now, and re-apply on top of the backports.
This should actually let us land a version of the patch much
closer to the version that will eventually land upstream.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:571567
TEST=build & boot to UI on oak-rev4
TEST=build & CQ test smaug, strago

Change-Id: Ia6d06ab3ff58412692e01a2fc2be35cbe8877ca4
Reviewed-on: https://chromium-review.googlesource.com/319742
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>