FROMLIST: ASoC: hdac_hdmi: Assign pin for stream based on dapm connection
Now that we have all the widgets enumerated we can route the stream
to any pin widget based on Mux connection. So map the pin to stream
accordingly.

Also seems the connection list to the pin widgets are not static
and is updated runtime based on type of display attached to the port.
This looks to be a hardware behavior.

So querying of the connection list is removed from pin initialization
and used in selecting the pin for a dai map.

Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
(am from https://patchwork.kernel.org/patch/7945851/ )
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>

BUG=chrome-os-partner:48269
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: Ie539c39efc6c35f3ce188b408603341f78ce98ee
Reviewed-on: https://chromium-review.googlesource.com/316207
Commit-Ready: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Tested-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>