BACKPORT: PM / Domains: add setter for dev.pm_domain
Adds a function that sets the pointer to dev_pm_domain in struct device
and that warns if the device has already finished probing. The reason
why we want to enforce that is because in the general case that can
cause problems and also that we can simplify code quite a bit if we can
always assume that.

This patch also changes all current code that directly sets the
dev.pm_domain pointer.

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Reviewed-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 989561de9b5112999475b406557d9c7e9e59c041)

Conflicts:
	drivers/acpi/acpi_lpss.c
	drivers/base/power/clock_ops.c
	drivers/gpu/vga/vga_switcheroo.c

BUG=chrome-os-partner:46491
TEST=suspend/resume on glados/chell

Change-Id: Icb236ec9ffbdcb7b57b1870fe7ecee066f4ac1ec
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/326881
Reviewed-by: Benson Leung <bleung@chromium.org>