UPSTREAM: drm/i915: Restore inhibiting the load of the default context
Following a GPU reset, we may leave the context in a poorly defined
state, and reloading from that context will leave the GPU flummoxed. For
secondary contexts, this will lead to that context being banned - but
currently it is also causing the default context to become banned,
leading to turmoil in the shared state.

This is a regression from

commit 6702cf16e0ba8b0129f5aa1b6609d4e9c70bc13b [v4.1]
Author: Ben Widawsky <benjamin.widawsky@intel.com>
Date:   Mon Mar 16 16:00:58 2015 +0000

    drm/i915: Initialize all contexts

which quietly introduced the removal of the MI_RESTORE_INHIBIT on the
default context.

v2: Mark the global default context as uninitialized on GPU reset so
that the context-local workarounds are reloaded upon re-enabling.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Michel Thierry <michel.thierry@intel.com>
Cc: Mika Kuoppala <mika.kuoppala@intel.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Link: http://patchwork.freedesktop.org/patch/msgid/1448630935-27377-1-git-send-email-chris@chris-wilson.co.uk
Reviewed-by: Mika Kuoppala <mika.kuoppala@intel.com>
Cc: stable@vger.kernel.org
[danvet: This seems to fix a gpu hand on after the first resume,
resulting in any future suspend operation failing with -EIO because
the gpu seems to be in a funky state. Somehow this patch fixes that.]
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 42f1cae8c079bcceb3cff079fddc3ff8852c788f
git://anongit.freedesktop.org/drm-intel drm-intel-nightly)
Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>

BUG=chrome-os-partner:48349
TEST=Ensure No graphic regressions found after system boots.
TEST=Ensure playing Video using HWA VP9 Decoding and launching camera
     same time doesn't cause the camera stream to get corrupted.

Change-Id: Ie625f2f91ea8897b5a19d070961ec7881eabf4e3
Reviewed-on: https://chromium-review.googlesource.com/321200
Commit-Ready: Marc Herbert <marc.herbert@intel.com>
Tested-by: Russ Sage <russ.sage@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>