CHROMIUM: clk: Fix out-of-bounds read in clk_notifier_register()
If the struct clk_notifier corresponding to the given struct clk is not
found in the global clk_notifier_list, cn will not point to a valid
struct clk_notifier.  Use a temporary to iterate over the list and only
set cn if the matching struct clk_notifier is found.

BUG=chromium:581371
TEST=Build and boot on Smaug.

Change-Id: Ic3c87656a4e37ca69422aac1a0881d49d8006921
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/323838
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Alexander Potapenko <glider@chromium.org>