CHROMIUM: ion: Fix use-after-free in ion_debug_pid_show()
If a process has an ion/clients/pids/<pid> debugfs file open when the
corresponding client is closed, a read() from the file may result in a
use-after-free since the associated struct ion_pid_data is freed in
ion_client_destroy().  Fix this by getting a reference to the struct
ion_pid_data in opne() and putting it on release().

Note that this is a general problem with debugfs - there is no good way
to tell in the read() patch whether or not the corresponding file has
been unlinked by debugfs_remove().

BUG=b:26270027
TEST=Build and boot on Smaug; read from /d/ion/clients/pids/.

Change-Id: I1a2026bb534db6417ffd0bd16a78d17fe89a2a73
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/324616
Reviewed-by: David Riley <davidriley@chromium.org>
(cherry picked from commit 6a598dfa9aca1998d4a7f2cfff0130138261ee11)
Reviewed-on: https://chromium-review.googlesource.com/327689