CHROMIUM: drm/nouveau: Don't reference vma after unmap
In nouveau_gem_ioctl_unmap(), after the call to nouveau_gem_object_unmap(),
assume that the vma can be freed at any point after that, and avoid further
references to the vma.  The only other reference to the vma after this point
is a peek at the as to see if the address space allocation is sparse, so
save off as before calling nouveau_gem_object_unmap().

BUG=chrome-os-partner:52570
TEST=Build and boot to desktop on A44.

Change-Id: I815dbd85f113f25e5f1a0cbd3117466e23d2cf1d
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/340515
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>