Revert "ecryptfs: forbid opening files without mmap handler"
(cherry picked from commit 78c4e172412de5d0456dc00d2b34050aa0b683b5)

This reverts commit 2f36db71009304b3f0b95afacd8eba1f9f046b87.

It fixed a local root exploit but also introduced a dependency on
the lower file system implementing an mmap operation just to open a file,
which is a bit of a heavy hammer.  The right fix is to have mmap depend
on the existence of the mmap handler instead.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Cc: stable@vger.kernel.org
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>
Change-Id: Icbcc37c20fac09eba86b44099001d683cd4cbdea
Bug: 29444228