CHROMIUM: drm/tegra: Not all tegra_bos have "->num_pages"
Patch "Check buffer size when doing kmap" adds a check at the
beginning of "tegra_bo_kmap" to make sure the request page
doesn't exceed the buffer size. But that patch does it in a wrong
way which is, not all "tegra_bo"s have "->num_pages" programmed.

Actually in current TegraDRM driver, we have 3 types of BOs:
- non-IOMMU bo, alloc/free using dma mapping api
- IOMMU bo which has non-contiguous physical pages backed
- buffers imported from dmabuf

Only IOMMU BOs have "->num_pages" because we indeed manally
allocate pages for the buffer.

So this patch instead using "bo->gem.size" to get the size
of the buffer then do the sanity check. The size of the GEM
object has correct value among the 3 types of BOs.

BUG=chrome-os-partner:59244
TEST=Video playback(Youtube) works and the POC app doesn't
     crash kernel anymore

Change-Id: Ifece387983bf45210b1be4b9e16920197b42bb78
Signed-off-by: Mark Zhang <markz@nvidia.com>