io: cros_ec: Move sensors preparation for suspend earlier
As we prepare to suspend sensors, we may get a sensor interrupt,
specially when the frequency set is very high. When going to suspend
this frequency is lowered during pm prepare phase to allow device to
suspend without waking up too soon. However before reaching this phase
there are other tasks in the suspend sequence (such as freezing
processes) that can take some time, and get a sensor interrupt before
reaching it and thus causing a wakeup. In order to prevent this from
happening, move the sensors preparation for suspend before other tasks
in order to allow it to go to suspend.

Signed-off-by: Adrian Salido <salidoa@google.com>
Bug: 25425420
Change-Id: I8c27c329d9e26c17dfb451eee20750f6693eb204